// apps/web/src/features/settings/types.ts
export interface CreateTeamInput {
  name: string;
  key: string;
  color?: string;
  description?: string;
  leadId?: string;
  memberIds?: string[];
}

export interface EditTeamInput {
  name: string;
  key: string;
  color?: string;
  description?: string;
  leadId?: string;
  memberIds?: string[];
}